Deprecate :: in record field names
Due to {ID} including ::, it is currently possible to create records
with fields that contain ::. This mostly seems to work, but I don't
think we should allow it. In fact, it is allowed in function args as well,
but ends-up failing trying to lookup the field name in a module rather
than in the local scope.